requirements.txt
fastapi
uvicorn
PyPDF2
opencv-python
numpy
Pillow
piexif
pydub
librosa
